Add search param fields
Adding support for the search param `fields`, https://developer.box.com/reference/get-search/#param-fields

From the docs:
fields `string array` in query optional
example
id,type,name
A comma-separated list of attributes to include in the response. This can be used to request fields that are not normally returned in a standard response.

Be aware that specifying this parameter will have the effect that none of the standard fields are returned in the response unless explicitly specified, instead only fields for the mini representation are returned, additional to the fields requested.